python/requirements.txt
fastapi
uvicorn[standard]
torch
transformers
pydantic>=2.7.0,<3.0.0
python-dotenv